requirements/base.txt
dj-database-url==2.1.0
django-modeladmin-reorder==0.3.1
django-tailwind==3.6.0
django-bleach==3.1.0
django-polymorphic==3.1.0
Django==4.2.4
httpx==0.24.1
psycopg2==2.9.1
python-decouple==3.8